ARROW-430: Improved version handling
This reintroduces `setuptools_scm` versioning for git clones and sdists/wheels. git-archives are handled by a separate chunk of code that will infer the version from the `pom.xml`

As in the Maven world, always the next to-be-released version is specified and Python PEP 440 development versions are based on the previous release, it used the most minimal pre-release version specified there.

I would suggest to keep the conda package versioning as it is currently, i.e. manually set it to 0.1.0postX.

Also: I would rather not parse the Maven XML but that is the most simple way currently to ensure that the versioning system is in a state where we can still make releases with the Maven release plugin.

Author: Uwe L. Korn <uwelk@xhochy.com>

Closes #248 from xhochy/ARROW-430 and squashes the following commits:

39753f8 [Uwe L. Korn] Infer version from java/pom.xml
05c44ea [Uwe L. Korn] Get rid of setuptools_scm_git_archive
14b8136 [Uwe L. Korn] Revert "ARROW-429: Revert ARROW-379 until git-archive issues are resolved"

Change-Id: I4f6d291e63b2518af47c2a81049aa24a38c92821