ARROW-360: C++: Add method to shrink PoolBuffer using realloc
Added no explicit unit test for this as I want to keep the freedom to
the allocator in the future to advise the PoolBuffer on an acceptable
minimal size. In some cases it might be worth it to occupy a whole
page.

Resizing to a smaller size is tested though, so we already have a unit test ensuring that this code runs smoothly.

Author: Uwe L. Korn <uwelk@xhochy.com>

Closes #272 from xhochy/ARROW-360 and squashes the following commits:

f4992ee [Uwe L. Korn] Adjust DCHECK for zero size arrays
040d3b4 [Uwe L. Korn] ARROW-360: C++: Add method to shrink PoolBuffer using realloc