ARROW-506: Java: Implement echo server for integration testing.
While implementing this, it became clear it made sense for the stream writer to
have an API to indicate EOS without closing the stream. The current message the
reader is expecting is a 4 byte size for the next batch. This patch proposes we
allow 0 as the size to indicate EOS.

Author: Nong Li <nongli@gmail.com>

Closes #295 from nongli/echo_server and squashes the following commits:

c115b02 [Nong Li] Add license header.
a3a50ca [Nong Li] ARROW-506: Java: Implement echo server for integration testing.