ARROW-822: [Python] StreamWriter Wrapper for Socket and File-like Objects without tell()
Added a wrapper for StreamWriter to implement the required tell() method so that python sockets and file-like objects can be used as sinks.  The tell() method will report the position by starting at 0 when the StreamWriter is created and incrementing by number of bytes after each write.

Added unittests that use local socket as the source/sink for streaming.

Author: Bryan Cutler <cutlerb@gmail.com>

Closes #569 from BryanCutler/pyarrow-stream-writer-socket-ARROW-822 and squashes the following commits:

6cdec4f [Bryan Cutler] Removed StreamWriter wrapper and put position handling in PyStreamWriter instead
2bd669f [Bryan Cutler] Added StreamSinkWrapper to ensure stream sink has tell() method, added unittest for StreamWriter and StreamReader over local socket