ARROW-1086: include additional pxd files during package build
Since this is purely a packaging change, I believe I would need to change`travis_script_python.sh` in order test it (build a wheel, set up a new conda environment and install the wheel, etc). Not sure if that's worth doing? Happy to take a stab if it is.

Author: Jacob Scott <jhscott@infer.com>

Closes #733 from snakescott/ARROW-1086 and squashes the following commits:

6c3e8598 [Jacob Scott] [ARROW-1086] Add more pxd files to package_data