ARROW-1310: [JAVA] revert changes made in ARROW-886
@elahrvivaz , @StevenMPhillips

Reverting the changes made for ARROW-886 -- https://github.com/apache/arrow/pull/591

(1) Don't explicitly reallocate the offsetVector in realloc() function of Variable Length Vectors.   If we call setSafe() on variable length vector, it will internally invoke setSafe() on the corresponding offsetVector and the latter function can decide whether to reallocate the offsetVector or not.

(2) Doing (1) will break the unit test added as part of PR 591 so we need to remove that as well.

Author: siddharth <siddharth@dremio.com>

Closes #937 from siddharthteotia/ARROW-1310 and squashes the following commits:

c5a2707c [siddharth] ARROW-1310: revert changes made in ARROW-886