ARROW-1416: Clarify memory layout documentation
This PR slightly reduces ambiguity in the array example for null bitmaps. The original example was left/right symmetric; this PR changes the example to break that symmetry. Asymmetry is important since readers who skip the byte endianness section could have interpreted the bitmap buffer in two distinct ways: left-to-right with an offset of 3 (wrong), or right-to-left with zero offset (correct).

Author: Fritz Obermeyer <fritz.obermeyer@gmail.com>

Closes #998 from fritzo/patch-1 and squashes the following commits:

af3dcbd0 [Fritz Obermeyer] Clarify memory layout documentation