ARROW-1359: [C++] Add flavor='spark' option to write_parquet that sanitizes schema field names
I also made the default for `use_deprecated_int96_timestamps` None so that we can distinguish between unspecified and explicitly False. In the event that the user passes `flavor='spark'`, this is enabled. Once Spark processes the int96 deprecation in the future, we can remove this part.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #1076 from wesm/ARROW-1359 and squashes the following commits:

8a60b66 [Wes McKinney] Use composition rather than inheritance
e3fa8ec [Wes McKinney] Add note about spark flavor to Sphinx docs
8159a51 [Wes McKinney] Add flavor='spark' option to write_parquet that sanitizes schema field names, turns on int96 timestamps