ARROW-1635: Add release management guide
I tried to write down all of the details that I've accumulated in my head (and probably missed some things). There are a lot of small details that are omitted, but it's clear there's plenty of opportunities to make things easier on the release manager or those working on post-source release binary package builds.

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #1185 from wesm/ARROW-1635 and squashes the following commits:

3dc2edb5 [Wes McKinney] Write up release management procedure from my perspective
9b7423f9 [Wes McKinney] Start release management markdown doc