ARROW-1676: [C++] Only pad null bitmap up to a factor of 8 bytes in Feather format
cc @rvernica

Due to Arrow's buffer padding, the validity bitmap may be larger than the Feather format expects it to be. This can result in off-by-one errors when reading files.

It's disappointing that this correctness issue has existed for so long. I am looking to determine if this might have affected users of the Python API

Author: Wes McKinney <wes.mckinney@twosigma.com>

Closes #1204 from wesm/ARROW-1676 and squashes the following commits:

60a2ed55 [Wes McKinney] Fuzz test more array lengths
540f79d3 [Wes McKinney] Ensure proper padding
fa677cc2 [Wes McKinney] Only write necessary bytes from null bitmap in Feather writer