ARROW-1826: [JAVA] Avoid branching in copyFrom for fixed width scalars
cc @jacques-n , @BryanCutler , @icexelloss

Note: there was no performance regression or different behavior introduced as part of refactor patches.

This is something I had noticed in the old code during refactoring but lost track of it. copyFrom function for fixed width scalars does a double check if the cell in the source vector is NULL or not. The old code used to do the same.

There is a straightforward way to avoid double check but the new changes avoid branch all-together; since it is fine to copy garbage from source vector as long as the validity bit is copied correctly to indicate NULL or non-NULL value in the target cell.

Will add unit tests.

Author: siddharth <siddharth@dremio.com>

Closes #1325 from siddharthteotia/ARROW-1826 and squashes the following commits:

8fc5b3f1 [siddharth] added unit tests
84de0cda [siddharth] ARROW-1826: [JAVA] Avoid branching in copyFrom