ARROW-3021: [Go] add support for List arrays
This CL introduces array.Builder, a generic interface to build
array.Interface.
It is needed to be able to build Lists' array values.

array.NewListArray and array.NewListBuilder are introduced to,
respectively, create a new List array value and a builder value to
populate the contents of a List array.

Author: Sebastien Binet <binet@cern.ch>

Closes #2402 from sbinet/issue-3021 and squashes the following commits:

255b425a <Sebastien Binet> ARROW-3021:  add support for List arrays