ARROW-2038: [Python] Strip s3:// scheme in S3FSWrapper isdir() and isfile()
This fixes an exception from ParquetDataset arising when the supplied path contains the `s3://` scheme specifier. The issue stemmed from the fact that while the underlying S3FileSystem does support both types of paths, with and without and explicit `s3://`, its function calls always return paths stripped of the scheme. This messed up with the logic in isdir() and isfile().

An alternative solution would be to strip the scheme in parquet.py (by adding it to _URI_STRIP_SCHEMES). This however would require additional code changes along the lines of:

```python
_URI_STRIP_SCHEMES = ('hdfs', 's3')

def _parse_uri(path):
    path = _stringify_path(path)
    parsed_uri = urlparse(path)
    if parsed_uri.scheme in _URI_STRIP_SCHEMES:
        scheme = '{0}://'.format(parsed_uri.scheme)
        path = parsed_uri.geturl().replace(scheme, '', 1)
        return path
    else:
        # ARROW-4073: On Windows returning the path with the scheme
        # stripped removes the drive letter, if any
        return path
```

Not sure if that would have any impact on handling HDFS. Therefore this patch proposes a safer, more localised approach, already used in other parts of S3FSWrapper.

Author: Dmitry Vukolov <dmvuko@gmail.com>

Closes #3286 from dvukolov/master and squashes the following commits:

8de916c5 <Dmitry Vukolov> Strip s3:// scheme in S3FSWrapper isdir() and isfile()