ARROW-4739: [Rust] LogicalPlan can now be passed to threads
I have been working on a PoC of parallel execution with partitioned data sources. This PR is for one small step towards that - the ability to pass logical plans to threads. For example, we often want to execute the same plan simultaneously across all partitions.

The change was mostly just replacing `Rc` with `Arc`.

Author: Andy Grove <andygrove73@gmail.com>

Closes #3798 from andygrove/ARROW-4739 and squashes the following commits:

88f27d2e <Andy Grove> Add test
8e12fb8d <Andy Grove> Logical plan uses Arc instead of Rc