ARROW-3550: [C++] use kUnknownNullCount for the default null_count argument
Previous default for this parameter in Array constructors was 0, which is not correct for all bitmaps the array might've been constructed with.

Sidecar: corrected `TypeTraits<NullType>` to reflect that it is parameter free and has a type_singleton.

Author: Benjamin Kietzman <bengilgit@gmail.com>

Closes #3781 from bkietz/ARROW-3550-array-constructor-kUnknownNullCount and squashes the following commits:

feaa3d36d <Benjamin Kietzman> add simple test for default null_count
fbe491572 <Benjamin Kietzman> revert deep copy of ArrayData
6ebaf37fa <Benjamin Kietzman> use kUnknownNullCount for the default null_count argument