ARROW-4940: [Rust] Enable warnings for missing docs, add docs in datafusion
This is the datafusion part of [ARROW-4683].

Adds a warn lint to the crate root for missing docs, and attempts to add
docs for anything missing at the time of writing.

Overall, I felt like I was delivering a weak product here. Most of the stuff that didn't have docs really didn't seem to need them. I'm not sure how much value is being added here. Notably, I actually opted to allow missing docs on one enum for aggregate function types since the only thing I could think to add in a docstring would be virtually identical to the name of the variant.

I'd welcome suggestions for rephrasing, etc.

[ARROW-4683]: https://issues.apache.org/jira/browse/ARROW-4683

Author: Owen Nelson <onelson@gmail.com>

Closes #3952 from onelson/missing-docs-datafusion and squashes the following commits:

05fe579d <Owen Nelson> ARROW-4940:  rephrase docs, fixup typos, catch-up after rebase.
164772de <Owen Nelson> ARROW-4940:  Enable warnings for missing docs, add docs in datafusion.