ARROW-5412: [Integration] Add Java option for netty reflection
After ARROW-3191, Java requires the property `io.netty.tryReflectionSetAccessible` to be set to `true` for JDK >= 9. This is already in the root POM, but causes integration tests to fail. This adds the property and an option to the Java command when running integration tests.

Author: Bryan Cutler <cutlerb@gmail.com>

Closes #4522 from BryanCutler/java-integration-netty-conf-ARROW-5412 and squashes the following commits:

25bec06 <Bryan Cutler> fix flake check
d6f0eee <Bryan Cutler> Added '-Dio.netty.tryReflectionSetAccessible=true' property to Java integration test command