ARROW-5082: [Python] Stop exporting copies of shared libraries in wheel
This reduces the size from 50mb to 28mb. I haven't tested if this breaks the OSX wheel. Note that the fix is brittle since currently (on Linux) it links with the full-versioned shared library binary. This works out due to the `-len(x)` sorting applied. A proper fix would be to keep the one linked (found via `ldd` or some other methods). I suspect that auditwheel and subsequent install & test will catch this if the contract were to change.

Author: François Saint-Jacques <fsaintjacques@gmail.com>

Closes #4577 from fsaintjacques/ARROW-5082-wheel-size and squashes the following commits:

15932d00b <François Saint-Jacques> ARROW-5082:  Stop exporting copies of shared libraries in wheel