ARROW-5632: [Doc] Basic instructions for using Xcode with Arrow
Added a new section on debugging with Xcode on macOS.  I am a bit wary of doing a step-by-step tutorial with screenshots as those tend to get quickly out of date as the Xcode app evolves, and I think there will always be better sources elsewhere for helping users ramp up on how to use Xcode.  The goal is to add enough to help someone familiar with C++ development in Xcode to get started working on Arrow.

cc: @nealrichardson or @xhochy: I recall you both have macs, could you try these instructions out and let me know if you run into any trouble?  Any advice on clarity, readability, etc would also be very welcome.  I'm not sure if I found the right location to add this so let me know if this fits in ok with the rest of the overall cpp doc.

Author: Hatem Helal <hhelal@mathworks.com>

Closes #4596 from hatemhelal/arrow-5632 and squashes the following commits:

8cf81edf7 <Hatem Helal> Basic instructions for using Xcode with Arrow